refactor(stage3): consolidate parse_title and related-papers ranking
Candidate B, slices 3-4.

parse_title: duplicated with different signatures (text-based vs path-based).
- woeai/wechat/review.py: parse_title(markdown_text) - the pure text version.
- markdown_to_rtd.py: imports it directly (already text-based).
- wechat_draft.py: parse_title(article_path) becomes a thin wrapper that
  reads the file then calls the pure version.

rank_against_target: the candidate_rank closure was byte-for-byte identical
in both scripts' related-papers builders.
- woeai/wechat/backlog.py: rank_against_target(paper, target) - takes target
  as a param instead of closing over it.
- Both scripts' related-papers builders now sort with it via lambda.

55 tests green.
